+++ pytbo/__init__.py
@@ -1,3 +1,21 @@
 # -*- coding: utf-8 -*-
 
-from .bare import *
+__title__     = 'pytbo'
+__version__   = '0.1.0'
+__build__     = 0x000100
+__author__    = 'Alessandro Costa'
+__license__   = 'Apache 2.0'
+__copyright__ = 'Copyright 2016 Alessandro Costa'
+
+from .bare import BareBot
+from .types import ( Audio, CallbackQuery, Chat, ChosenInlineResult, Contact, Document, File, ForceReply,
+                     InlineKeyboardButton, InlineKeyboardMarkup, InlineQuery, InlineQueryResultArticle,
+                     InlineQueryResultAudio, InlineQueryResultCachedAudio, InlineQueryResultCachedDocument,
+                     InlineQueryResultCachedGif, InlineQueryResultCachedMpeg4Gif, InlineQueryResultCachedPhoto,
+                     InlineQueryResultCachedVideo, InlineQueryResultCachedVoice, InlineQueryResultContact,
+                     InlineQueryResultDocument, InlineQueryResultGif, InlineQueryResultLocation,
+                     InlineQueryResultMpeg4Gif, InlineQueryResultPhoto, InlineQueryResultVenue,
+                     InlineQueryResultVideo, InlineQueryResultVoice, InputContactMessageContent,
+                     InputLocationMessageContent, InputTextMessageContent, InputVenueMessageContent,
+                     KeyboardButton, Location, Message, MessageEntity, PhotoSize, ReplyKeyboardHide,
+                     ReplyKeyboardMarkup, Sticker, Update, User, UserProfilePhotos, Venue, Video, Voice )