+++ HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ---------------
 
+0.1.0 (2016-04-23)
+++++++++++++++++++
+
+* BareBot class that wraps Telegram Bot APIs methods
+* Types classes that wraps Telegram Bot APIs types
+
 0.0.1 (2016-04-14)
 ++++++++++++++++++